Teleport use Point2 instead Position, binds added, less spaghetti code in Main.java
