+++ build.gradle
@@ -27,7 +27,7 @@ repositories {
 }
 
 ext {
-    mindustryVersion = 'v156.1'
+    mindustryVersion = 'v157.3'
     jabelVersion = '1.0.1'
     isWindows = System.getProperty("os.name").toLowerCase().contains("windows")
     sdkRoot = System.getenv("ANDROID_HOME") ?: System.getenv("ANDROID_SDK_ROOT")
