+++ src/resources/bundles/bundle.properties
@@ -59,7 +59,7 @@ setting.check4update.description = Choose whether if the mod should check for up
 setting.subtitle.name = About Me
 setting.subtitle.description = This information will be visible to other users of this mod on servers with integration.\nI can't do anything with what you will write here, so I'll just ask to be cultured people.
 
-category.mod.name = Modification
+category.mod.name = Scheme SIze
 category.bt.name = Building Tools
 category.add.name = Additional
 category.radius.name = Radii
@@ -174,6 +174,16 @@ admins.notavailable = [scarlet]\u26A0[]This feature is unavailable because you a
 admins.notsupported = [scarlet]\u26A0[]This feature is not supported by the darkdustry plugin!
 admins.nocore = [scarlet]\u26A0[]There is no core in the selected team!
 admins.nounit = [scarlet]\u26A0[]The selected team cannot create more units!
+
+keybind.scheme_teleport.name = Teleport
+keybind.scheme_despawn.name = Unit despawn
+keybind.scheme_team.name= Team menu
+keybind.scheme_ai.name = AI menu
+keybind.scheme_core.name = Core place
+keybind.scheme_unit.name= Unit menu
+keybind.scheme_unit_spawn.name = Units spawn menu
+keybind.scheme_effect.name = Effect menu
+keybind.scheme_item.name = Item menu
 
 console.classic = Classic
 console.multiline = Multiline
